Issue #1043: Single product import via API
- Move single product import via api in own test class
- add test whether product was imported into data pool
