+++ tests/Integration/Suites/CatalogImportApiTest.php
@@ -86,47 +86,4 @@ class CatalogImportApiTest extends AbstractIntegrationTest
         $this->assertSame(202, $response->getStatusCode());
         $this->assertSame('', $response->getBody());
     }
-
-    public function testV1ProductImportApiV1PutRequestHandler()
-    {
-        $testDataVersionString = 'foo-123';
-        $httpUrl = HttpUrl::fromString('http://example.com/api/product_import');
-        $httpHeaders = HttpHeaders::fromArray([
-            'Accept' => 'application/vnd.lizards-and-pumpkins.product_import.v1+json',
-        ]);
-
-        $productJson = json_encode([
-            'sku'        => 'led-arm-signallampe',
-            'type'       => 'simple',
-            'tax_class'  => 'class-1',
-            'attributes' => [
-                'backorders'  => true,
-                'url_key'     => 'led-arm-signallampe',
-                'description' => 'LED Arm-Signallampe<br />
-<br />
-LED Arm-Signallampe mit elastischem Band und Flasher mit variabler Blinkfolge,
-Flasher abnehmbar.',
-            ],
-        ]);
-
-        $httpRequestBodyString = json_encode(['product_data' => $productJson, 'data_version' => $testDataVersionString]);
-        $httpRequestBody = new HttpRequestBody($httpRequestBodyString);
-        $request = HttpRequest::fromParameters(HttpRequest::METHOD_PUT, $httpUrl, $httpHeaders, $httpRequestBody);
-
-        $factory = $this->prepareIntegrationTestMasterFactoryForRequest($request);
-        $implementationSpecificFactory = $this->getIntegrationTestFactory($factory);
-
-        $commandQueue = $factory->getCommandMessageQueue();
-        $this->assertEquals(0, $commandQueue->count());
-
-        $website = new InjectableDefaultWebFront($request, $factory, $implementationSpecificFactory);
-        $response = $website->processRequest();
-
-        $message = $this->getNextMessageFromQueue($commandQueue);
-        $this->assertSame(UpdateProductCommand::CODE, $message->getName());
-        $this->assertSame($testDataVersionString, $message->getMetadata()['data_version']);
-
-        $this->assertSame(202, $response->getStatusCode());
-        $this->assertSame('', $response->getBody());
-    }
 }
